+++ src/main/resources/kb/zango/domain/board/mapper/BoardMapper.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="kb.zango.domain.diary.mapper.BoardMapper">
+<mapper namespace="kb.zango.domain.board.mapper.BoardMapper">
     <insert id="insertBoard" parameterType="kb.zango.domain.board.entity.Board"
             useGeneratedKeys="true" keyProperty="boardId">
         INSERT INTO board (user_id, title, content, like_cnt, comment_cnt, regi_date, sc_id, board_type)
